refactor: move settings logic to module
